[refactor] 브레드크럼과 매뉴얼 버튼 컴포넌트로 분리
