+++ libs/components-shared/src/lib/MypageLayout.tsx
@@ -1,4 +1,5 @@
-import { Box, Flex, Stack, Link, Button } from '@chakra-ui/react';
+import { Box, Flex } from '@chakra-ui/react';
+import { UserType } from '@prisma/client';
 import {
   broadcasterCenterMypageNavLinks,
   MypageLink,
@@ -11,46 +12,20 @@ import {
   useCloseLiveShoppingStateBoardIfNotLoggedIn,
   useDisplaySize,
   useIsLoggedIn,
-  useManualLinkPageId,
 } from '@project-lc/hooks';
-import { useRouter } from 'next/router';
 import React from 'react';
-import NextLink from 'next/link';
-import { UserType } from '@prisma/client';
 import { FloatingHelpButton } from './FloatingHelpButton';
-import MypageBreadcrumb from './MypageBreadCrumb';
+import { MypageToolbar } from './MypageToolbar';
 import { Navbar } from './Navbar';
 import DesktopMypageSidebar from './navbar/DesktopMypageSidebar';
 
 const NAVBAR_HEIGHT = 67;
 const FOOTER_HEIGHT = 60;
 
-interface MypageLayoutProps {
+export interface MypageLayoutProps {
   children: React.ReactNode;
   appType?: UserType;
   navLinks?: Array<MypageLink>;
-}
-
-/** 해당 페이지 routerPath에 맞는 이용안내로 이동하는 링크버튼 */
-export function ManualLinkButton({
-  userType,
-}: {
-  userType: UserType;
-}): JSX.Element | null {
-  const router = useRouter();
-  const { data: manualId } = useManualLinkPageId({
-    routerPath: router.pathname,
-    userType,
-  });
-
-  if (!manualId) return null;
-  return (
-    <NextLink href={`/mypage/manual/${manualId}`} passHref>
-      <Button as={Link} size="sm" isExternal colorScheme="blue">
-        이용안내
-      </Button>
-    </NextLink>
-  );
 }
 
 export function MypageLayout({
@@ -99,11 +74,7 @@ export function MypageLayout({
               className="content-wrapper"
               minHeight={`calc(100vh - ${NAVBAR_HEIGHT}px - ${FOOTER_HEIGHT}px)`}
             >
-              <Stack direction="row" spacing={4} alignItems="center">
-                <MypageBreadcrumb />
-                <ManualLinkButton userType={appType} />
-              </Stack>
-
+              <MypageToolbar appType={appType} />
               {children}
             </Box>
             {/* 하단 푸터 */}
